requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
defusedxml==0.7.1
    # via -r requirements.in
django==5.1
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==4.7.0
    # via -r requirements.in
django-nested-admin==4.1.1
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.28.0
    # via -r requirements.in
et-xmlfile==2.0.0
    # via openpyxl
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
numpy==2.2.2
    # via pandas
openpyxl==3.1.5
    # via -r requirements.in
pandas==2.2.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-monkey-business==1.1.0
    # via django-nested-admin
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via -r requirements.in
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.38
    # via -r requirements.in
sqlparse==0.5.3
    # via django
typing-extensions==4.12.2
    # via
    #   referencing
    #   sqlalchemy
tzdata==2025.1
    # via pandas
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.3.0
    # via requests
